requirements.txt
tqdm
numpy~=1.20.3
pillow~=8.3.1
cython~=0.29.22
opencv-python~=4.5.3.56
matplotlib~=3.2.2
shapely~=1.8.1
geopandas~=0.10.2
pandas~=1.4.1
pyproj~=3.3.0
setuptools~=47.3.1
boto3~=1.14.7
torch>=1.7.1
torchvision>=0.8.2
tensorboard>=2.2
wandb>=0.12.0
sentinelhub~=3.4.4
eo-learn-core~=1.0.0
eo-learn-io~=1.0.0
eo-learn-geometry~=1.0.0
eo-learn-visualization~=1.0.0
